Support systemd and ceph versions >= 9.2
Currently the devstack ceph plugin uses installation,
configuration, startup, and shutdown commands that do not always
work correctly for Ceph releases after hammer or with systemd
init system.

Check the ceph version and the init system and use appropriate
commands.

Co-Authored-By: Ramana Raja <rraja@redhat.com>

This is the same as https://review.openstack.org/#/c/332484/11